Co-authored-by: Daniel Xu <accounts@dxuuu.xyz>
Added a sample application for libbpf-rs. It will output all TCPv4
connections that have been established on the host as ips and
hostnames by probing `tcp_v4_connect` in kernel and glibc's
`getaddrinfo` in userland. On a successful host lookup the first
result will be stored in a hashmap, which can be used as a lookup
table to retrieve a hostname for ip_v4 connections.

When the program is started there should be console output when
TCPv4 connections are established on the system, eg. via curl:

host event: www.jsonplaceholder.com
ip event: 172.67.201.157