Avoid profiling offline or not present CPUs
Running example/c/profile on a virtual machine results in
libbpf_num_possible_cpus calculating the wrong number of CPUs, causing the
program to fail to set up performance monitor on a CPU/Core and throw an
error. In contrast, the Rust implementation ignores this error and continues
to execute.

This fix only applies to the C implementation.

Signed-off-by: ZhengSong Tu <tzssangglass@gmail.com>