rust/examples/profile: Handle errors more gracefully
It is entirely possible for symbolization to fail for legitimate reasons
when profiling the system continuously. For example, a process could
terminate between the time addresses from it were captured and we
attempt to symbolize them, in which case an error will be reported by
blazesym.
As it stands such an error will shoot down the profiler, which is not
the desired behavior. With this change we handle such errors more
gracefully by just printing an error message.

Signed-off-by: Daniel Müller <deso@posteo.net>