examples/rust: Fix improper error handling
The error handling in the profile example is insufficient. Specifically,
we never check for the success of perf_event_open. If it fails, we end
up passing an FD of -1 to libbpf's bpf_program__attach_perf_event_opts,
which in turns barks at us.
Make sure to properly check for errors. Also address clippy issues and
improper formatting along the way.

Signed-off-by: Daniel Müller <deso@posteo.net>