examples/rust: Store skeletons in $OUT_DIR
Store skeletons in $OUT_DIR as opposed to .output, which is kind of the
proper thing to do (though it certainly doesn't help
discoverability...) and it is more in line with how examples in
libbpf-rs work. Also remove the "there is hope!" comments, because even
with https://github.com/rust-lang/rust/pull/83366 resolved we still
cannot use the concat! macro in #[path = ...] attributes. All hope is
lost.

Signed-off-by: Daniel Müller <deso@posteo.net>